Separate index from the main process
